+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 **1.2.1**
 
   - fix `display:block` added to error elements (#230) (PR #413)
+  - added some more translations
+  - added `data-parsley-namespace` feature to allow W3C compliant DOM-API (#424)
 
 **1.2.0 (current stable)**